TL;DR: Added unit tests, refactored Simulator and fixed RLEParser bug.
Created unit tests for GameBoard and Simulator to ease our further work on dynamic game boards and threaded simulators.

The simulator is now an abstract class which supports one single SimRule, as opposed to the previous interface implementation with multiple rules. Its constructor requires a rule to be set at instantiation. This makes a simpler simulator with better performance  in terms of execution speed. Support for multiple rules could always be added back in later through a specialized SimRule implementation.

Fixed a bug in the RLE parser where numbers in front of '$' (end of line) got carried over to the next row. Turns out this number specifies the amount of empty rows to be followed.